requirements.txt
Django>=4.2
whitenoise
python-dotenv
